labs/07-face/python/detect-people/requirements.txt
azure-ai-vision-imageanalysis
azure-cognitiveservices-vision-customvision
azure-core
debugpy
matplotlib
pillow
python-dotenv